// server/src/types/PracticeSettings.ts
export type PracticeSettings = {
  practiceName: string
  website: string
  visitAddress: string
  postalAddress: string
  postalCodeCity: string
  contactName: string
  contactRole: string
  contactPhone: string
  contactEmail: string
  tintColor: string
  logoDataUrl: string | null
  updatedAtUnixMs: number
}